requirements.txt
# Shan-D Ultra-Human AI Requirements 📃
# Created by: ◉Ɗєиνιℓ 🧑‍💼
# Production-ready dependencies for maximum human-like AI

# Core dependencies
python-telegram-bot>=20.7
aiohttp>=3.9.0
asyncio>=3.4.3

# AI Model APIs
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Image/Media processing
Pillow>=10.0.0
aiofiles>=23.0.0

# System monitoring
psutil>=5.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Utilities
pydantic>=2.0.0
dataclasses-json>=0.6.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0

#Security
bandit
safety

#webhook
fastapi
httpx
